Move the code that spawns the file command to a separate file.
