+++ lib/index.js
@@ -7,7 +7,7 @@ var util = require('util'),
     fs = Promise.promisifyAll(require('fs')),
     os = require('os'),
     constants = process.ENOENT ? process : require('constants'),
-    childProcess = require('child_process'),
+    determineFileType = require('./util/determineFileType'),
     _ = require('lodash'),
     EventEmitter = require('events').EventEmitter,
     pathModule = require('path'),
@@ -529,24 +529,11 @@ _.extend(AssetGraph.prototype, {
             } else if (rawSrc.length === 0) {
                 return foundType(); // Give up
             } else {
-                // Work the magic
-                var fileProcess = childProcess.spawn('file', ['-b', '--mime-type', '-']),
-                    fileOutput = '';
-
-                fileProcess.on('error', function (err) {
+                return determineFileType(rawSrc).then(function (contentType) {
+                    return foundType(AssetGraph.lookupContentType(contentType));
+                }, function (err) {
                     that.emit('warn', 'Could not determine asset type from contents using the file command: ' + err.stack);
-                });
-
-                // The 'file' utility might close its stdin as soon as it has figured out the content type:
-                fileProcess.stdin.on('error', function () {});
-                fileProcess.stdin.end(rawSrc);
-
-                return new Promise(function (resolve, reject) {
-                    fileProcess.stdout.on('data', function (chunk) {
-                        fileOutput += chunk;
-                    }).on('end', function () {
-                        resolve(foundType(AssetGraph.lookupContentType(fileOutput.match(/^([^\n]*)/)[1])));
-                    });
+                    return foundType();
                 });
             }
         })
